api/requirements.txt
Django>=4.2,<5.0
django-cors-headers>=4.4.0
django-distill>=3.2.7
dj-database-url>=2.1.0
psycopg2-binary>=2.9.10
gunicorn>=20.1.0
strawberry-graphql[django]>=0.209.0
pydantic>=2.9.0
requests>=2.32.3
redis>=5.0.8
django-redis>=5.4.0
python-dotenv>=1.0.0
croniter>=2.0.7
PyYAML>=6.0.2
whitenoise>=6.0.0
